allocation library updates
Summary:

Relaxed the transitional invariant to allow the counterpart of Revoked/Allocated to be unassigned. That improves
its usability where we want to release/revoke live work in various conditions, but don't want to deal with an assignment. Added load balancing. Removed domain parameter -- it does not pull its weight the way we'd use it (Splitter coordinator only, but there an internal anti-affinity rule seems more suitable).

Updated Splitter leader

GitOrigin-RevId: b0425110fd8d8a368583fa0b6f79d2fa3d21d2e4