limit Cluster interface and expand resolver
Summary:

Made Cluster read-only and limited to what user need. Updated the lookup to search by a list of grant states to support different resolution behavior. We will need that if and when users want to handle transitional states.

Workpool and coordinator use the cluster implementation struct directly, which has more methods than the interface. Made the workpool emit a copy of the cluster to make it safe and immutable.

GitOrigin-RevId: c261285b05bef0ad471187453284fef25f97d931