Update stale grants on re-connect
When a consumer is re-connected with stale leases, their states are not updated. For example, a consumer was assigned an allocated grant, but disconnected before receiving promotion for that grant; it may re-connect after the grant was promoted. The server may send an active grant in the assignment, but local state of the grant is not updated - it continues to stay allocated and doesn't notify handles about the update.

With this change grants received after re-connect are used to update the state of the grant.

GitOrigin-RevId: 1a066fab719c1df5182a20aa7799b1fca5d8f2f7