Fix grant expiration tracking
Grant expiration is stored in multiple places: original grant received from the server, the lease property of `grant` and the expiration field of `ownership`. This adds unnecessary complexity when updating the expiration. For example, right now `extend` updates informational expiration (in `ownership`) for all grants, including revoked.

The change modifies `ownership` to use the lease from `grant` to get the expiration. This removes the need to update individual leases on extend.

GitOrigin-RevId: db09a929d6c47f38ab293375ed79f2e955574826