Send shards in cluster messages from coordinator
This change updates Splitter coordinator to send a list of valid shards in cluster messages:

- every time, in cluster snapshot: to initialize a list of valid shards on a client,
- when list of shards has changed, in cluster change message: this is to update clients when some domain has changed its shards (e.g. resharding, placement update, removal or addition of a domain).

Note that the coordinator sends the whole list of shards whenever it decides to send them, not as an incremental update. This can be quite large for some clients , but it's not expected to happen often.

The client change to handle shards in the cluster snapshot is already implemented in the earlier change. The cluster update logic is extended to support the new "Shards" message that contains a list of shards. Note that this list can be empty, indicating that there are no shards in a service.

The cluster update logic doesn't validate current grants to use valid shards. This is temporary omission while the service is not sending all valid shards in the initial snapshot and there is a possibility that the list of valid shards is not complete (when server doesn't send a list of shards, it's computed from current grants).

Cluster tests will be updated with the following change, where shards will be used to invalidate grants assigned for unknown shards.

GitOrigin-RevId: f8539625de4412f57db9fc33efd60681b63eb7e5