Revoke grants locally when workpool draining and disconnected
When Workpool is draining it will no longer re-connect to the server. If during that time it disconnects from the coordinator, then the assigned grants can be revoked locally because there won't be any messages from the coordinator. By doing that it signals the client that the work needs to be finished.

<REDACTED>

GitOrigin-RevId: 9e681ee4f8bb68ca44fa5dc362afd41d604ca4c6