Validate cluster update messages
This change introduces validation for the cluster messages. It performs a sanity check to make sure a coordinator sent a valid message. When an error is detected, a workpool disconnects from the coordinator.

The validation checks the following:

- duplicates: grant IDs, consumer IDs, shards
- all grants have valid shards
- grants assigned to the same shard are compatible
- retained grants and consumers cannot be used in "change" messages before they are assigned (either in a snapshot or assignment).

Validation is performed on each messages, but its enforcement is disabled by default so that existing clients wouldn't disconnect from Splitter if there is a bug in the validation logic. It will be enabled by default after some time.

GitOrigin-RevId: 693e77744306848c6ea200a13ddf9fb1e02ccb58