Fix cluster change bug with updated grants
The issue happens when two grants for the same shard survive a snapshot, but only allocated is known to the coordinator (when it sends a cluster change). When activating the allocated grant, an update may not be applied by a consumer, resulting in invalid cluster map and inability to forward requests.

Also fixing an issue when validation logic doesn't remove unassigned grants. When a cluster update unassigns a grant and assigns a new grant to the same shard, the validation fails.

GitOrigin-RevId: 5842bb39587baae39b6fbece86f116c117959c88