Fix cluster consumer version not updated correctly
This is a minor issue causing a cluster to keep an invalid version for retained consumers. This does not result is any issues because the retention check is comparing the version to zero to determine if a consumer is retained. The retention check for consumers is used during validation check to decide whether grants of the consumer are known or not.

GitOrigin-RevId: b2e14499e12338933f50486b262ae3872a8a85e3