loadbalance should assign work to underloaded workers
Summary:
Tweak the load balance logic to find underloaded workers (below average load per worker) and find any eligible work for it.

Note that a unit domain is prone to move, as the algorithm favors moving a heavy load and sometimes pays the placement penalty to reduce the load on an overloaded worker. This PR adds a unit test to describe the issue but doesn't address it yet.

GitOrigin-RevId: c21512525a2e8caddf0837dcf8f6d1b8cf975aaf