[Splitter Clients] Add metric for requests handled locally
This change adds a metric for requests handled locally by a client. This metric can be used to detect issues with local processing.

The change involves adding a separate interface for a resolver that allows to determine a qualified domain key for a generic key. The QDK is used when recording the metric.

Splitter-internal forwarding and handling is refactored into separate methods because it's using a qualified service name. Splitter clients are not supposed to use such forwarding, and so the logic is copied to `core`.

GitOrigin-RevId: 8d8f8a1c7c8072703fe436adef0714bd6c22b9ba