[Go Client] Reorganize forwarding error handling
The purpose of this change is to separate error handling in Splitter forwarding logic from the Splitter client error handling.

Forwarding error handling is intended to be integrated into consumer error handling and it should not include Splitter client handling. Currently this coupling (by calling `UnwrapError` in `RetryOwnership1`) doesn't affect consumer services because they don't use Splitter `RetryOwnership1` (they use their own copies of this method) and only ownership errors are important for `RetryOwnership1` (other are ignored). After this change client can directly use `RetryOwnership1` if their errors are organized according to the recommendations in the comments.

This change moves `ErrNotOwned` and `ErrDraining` to `proxy` to indicate that these errors are designated for forwarding.

Added helper methods to convert these two errors to and from gRPC errors.

Renamed `WrapError`/`UnwrapError` to `AsGRPCError`/`FromGRPCError` to be concise - it's hard to remember whether the wrapped error is gRPC or the other way around.

GitOrigin-RevId: 7e01a201527afa783f8077c18b0813ac3f8387d9