[Go Client] Replace ErrNotFound with ErrNotOwned in forwarding
This change removes the usage of `ErrNotFound` from the forwarding logic. This error makes ownership resolution more complicated and doesn't give us any benefits. On the contrary, the forwarding doesn't retry in cases when cluster information is not available: `Resolve` returns `ErrNotFound` and `IsOwnershipError` doesn't check for `ErrNotFound`.

None of the callers of this method check errors or check for `ErrNotOwned`.

GitOrigin-RevId: d17bef425ec7fec412f5f7d4322424f8e0915148