Fix Allocation.Modify to update the grant modifier
`w.live` and `w.revoked` store value types (grants). `Modify` updates the grants properties (`mod` and `Mod`), but the grants in the maps are not updated.

This bug doesn't cause any issues because the notification sent by the coordinator contains the correct state for the updated grant. The grant state in the allocation is invalid, but it's not causing issues because the mod field affects only auxiliary states (`Loaded`, `Unloaded`) which are treated similar to the main states (`Allocated`, `Revoked`).

This change will lead to more frequent cluster map messages sent to consumers (that use Loaded and Unloaded states).

This change also adds logging to catch updates with invalid states.

GitOrigin-RevId: d0855566456c3099ca766579293628330a786d8f