Add Ownership to RangeFactory
Summary:
With [PR-97541](<REDACTED>), coordinator accepts revoke grant request initiated by the consumer.

To be able to request rovoke grant, the consumer need a mechanism to notify processor that the grant needs to be revoked. Ownership holds grant state information and could be used for such purpose.

This PR modifies RangeFactory to accept Ownership parameter, which could be passed to new `Range`. It is up to `Range` implementation if it request revoke in case of range restart.

This change is safe as it only changes RangeFactory signature without touching any code logic, any issue will be caught at compilation time instead of runtime.

GitOrigin-RevId: 36d1f73b554f8e3bf7c07c356550f19d6f8a289b